Several refactorings in BibTeXObject and its descendants classes.
Inclusion of the Visitor pattern to allow the implementation
of different formatters classes that define how each
BibTeXObject will be formatted.

Firstly, the class BibTeXObject was changed to an interface
to define a Visitable object. Other interfaces SingleValueBibTeXObject
and KeyValueBibTeXObject were introduced. They extend the BibTeXObject.

The classes BibTeXComment, BibTeXPreamble and BibTeXInclude now implement
the SingleValueBibTeXObject interface.
The class BibTeXString implements the KeyValueBibTeXObject interface.

The class BibTeXFormatter now implements the new BibTeXObjectVisitor interface,
being a Visitor object.
This class is the only existing formatter implemented and, in the future,
it may be renamed to become more meaningful.
It has the attributes that define the style of the formatting.
As it implements the Visitor pattern, for each different kind
of BibTeXObject, there is a specific visit method
that defines how this object have to be formatted.
By this way, the if's chains that were statically checking each
object type were removed. Now, a polymorphic and dynamic
approach is implemented with the Visitor pattern.
If a new kind of BibTeXObject is introduced,
we only need to add a new visit method to the interface
BibTeXObjectVisitor and the compiler will complain
the absence of the new method in the concrete classes.

It was implemented a Builder pattern to build
different Visitor objects (formatters)
in a fluent way.
Currently, the BibTeXFormatterBuilder has
two methods: one for the standard bibtex formating
and another one for a more concise one (that doesn't inserts spaces
around the equal signal, and doesn't put the closing bracket into a new line).

The classe BibTeXFormatter was refactored to extract a lot of methods for:
make each method smaller and trying to assign a single responsability for each one;
remove duplicated code; give more meaningful names for the methods.

This class uses the new BibTeXStringBuilder classe that has
some utility methods to construct a java StringBuilder object
containing the representation of a BibTeXObject.
This class is not to construct BibTeXString objects,
but a string representation of any BibTeXObject.
But I confess that I didn't find a better name to avoid confusion :(.